DatabaseCache class and usage
Created a class which manages a database cache. The cache is a json file which
holds database index names as keys, and their paths as vals. This cache is
updated whenever DatabaseCache's set_entry or remove_entry are called, and it
exposes entries read from the cache (if any) via get_entries. Takes a construct
only property which indicates the directory in which its cache should live

The server updates the cache whenever a PUT or DELETE request is handled
successfully, and on startup it creates any databases stored in the cache (or
removes entries for databases which no longer exist)

[endlessm/eos-sdk#1166]