Change xapian-bridge architecture
This commit changes the architecture of the xapian-bridge so that a path
to the database is always a mandatory parameter for the query.
PUT and DELETE methods are then not required anymore, as is the simple
GET with no specified query.

The goal is to remove a lot of the ad-hoc API from xapian-bridge and
making it simpler and effectively stateless. More API will be removed in
following commits.