database-manager: major stemming refactor
We should base the stemmer language on the "lang" parameter sent along with
every query, as opposed to only setting it up in the first query and persisting
the stemmer across the lifetime of the server.

Also, we no longer support the metadatabases that required prefixes to be
grouped by language codes, so remove the unused instances of that parameter
from various structs/methods.

[endlessm/eos-sdk#2660]