database-manager: add support for flatpak bundles
When a flatpak bundle instructs the bridge to read a database located
inside the sandbox, it will have a prefix of /app.
The bridge runs outside of the sandbox though, so it does not have
access to that path.

For this purpose, we introduce the query parameters "flatpak_app_id",
"flatpak_path", and "flatpak_manifest_path". Flatpak bundles must use
these instead of "path" and "manifest_path". When these parameters are
used, we use the Flatpak library to fetch the real path.

(Cosimo / continued by Philip)

https://phabricator.endlessm.com/T11772