Avoid unsigned integer overflow for query limits
A query may receive a "limit" parameter to limit the number of results
in the matches set. The parameter is stored as a string, and parsed out
as a double precision floating point value, which then gets turned into
an unsigned integer through C casting.

Sadly, the "limit" parameter can contain a negative value — usually as a
shortcut meaning "give me all the results". This gets parsed into a
negative double, and then — in theory — it would overflow the unsigned
integer and be turned into `(uint) -1`, or `UINT_MAX`.

This chain of operations relies on various assumptions that may not be
true; while unsigned integers will wrap with a defined behavior, the
implicit conversion from a negative double to unsigned integer is
*undefined*:

| When a finite value of real floating type is converted to an integer
| type other than _Bool, the fractional part is discarded (i.e., the value
| is truncated toward zero). If the value of the integral part cannot be
| represented by the integer type, the behavior is undefined.

In order to get into defined behavior, we should ensure that the
floating point value we have is within the allowed ranges; we can keep
the "-1 means all results" shortcut to avoid having to change every
single caller.

https://phabricator.endlessm.com/T15009