Add support for passing filters separately
Filters can now be specified with "filter" and excluding filters with
"filterOut".  This means that "q" can be used for just the user's query,
which allows various built-in Xapian query parser behaviour to work more
naturally.

It also eliminates unnecessary use the less efficient "pure NOT" when
there are excluding filters.

New feature flag "query-param-filter" indicates that these new
parameters are supported.