Fix database_is_empty() code to match docs
The documentation comment say "Checks if the given database is empty
(has no documents)" but the code was actually checking if get_uuid()
returned an empty string, which essentially means either there are no
subdatabases or the backend doesn't support UUIDs.  There could be
subdatabases which are empty.

The result of this check is used to short-cut and return an empty
result set, so fixing the code to match the documentation makes
most sense.